Split client and querybuilder into separate files
